refactor: Upgrade PHPStan to level 9

+++ app/Actions/Fonts/DeleteFontsAction.php
@@ -8,6 +8,10 @@ final readonly class DeleteFontsAction
 {
     public function execute(string $filename): void
     {
+        if (empty($filename)) {
+            return;
+        }
+
         if (Storage::disk('local')->exists('fonts/' . $filename)) {
             Storage::disk('local')->delete('fonts/' . $filename);
         }
